refactor(core-lib): move mutler file handler to @musica/core
